Move offer, bondTradeTransactions, fraktBond schemas to shared
