+++ src/api/nft/core/schemas.ts
@@ -1,17 +1,13 @@
-import {
-  BondFeatures,
-  BondTradeTransactionV2State,
-  BondTradeTransactionV2Type,
-  BondingCurveType,
-  FraktBondState,
-  LendingTokenType,
-  PairState,
-  RedeemResult,
-  RepayDestination,
-} from 'fbonds-core/lib/fbond-protocol/types'
 import { z } from 'zod'
 
-import { NFTSchemaNew, StringIntSchema, StringPublicKeySchema } from '@banx/api/shared'
+import {
+  BondTradeTransactionSchema,
+  FraktBondSchema,
+  NFTSchemaNew,
+  OfferSchema,
+  StringIntSchema,
+  StringPublicKeySchema,
+} from '@banx/api/shared'
 
 const MarketMetaSchema = z.object({
   marketApr: StringIntSchema,
@@ -33,95 +29,6 @@ export const MarketPreviewSchema = z
     tensorSlug: z.string().optional(), //TODO Remove when BE fix this
   })
   .merge(MarketMetaSchema)
-
-const BondingCurveSchema = z.object({
-  delta: StringIntSchema,
-  bondingType: z.nativeEnum(BondingCurveType),
-})
-
-const ValidationPairSchema = z.object({
-  loanToValueFilter: StringIntSchema,
-  collateralsPerToken: StringIntSchema,
-  maxReturnAmountFilter: StringIntSchema,
-  bondFeatures: z.nativeEnum(BondFeatures),
-})
-
-export const OfferSchema = z.object({
-  publicKey: StringPublicKeySchema,
-  assetReceiver: StringPublicKeySchema,
-  baseSpotPrice: StringIntSchema,
-  bidCap: StringIntSchema,
-  bidSettlement: StringIntSchema,
-  bondingCurve: BondingCurveSchema,
-  buyOrdersQuantity: StringIntSchema,
-  concentrationIndex: StringIntSchema,
-  currentSpotPrice: StringIntSchema,
-  edgeSettlement: StringIntSchema,
-  fundsSolOrTokenBalance: StringIntSchema,
-  hadoMarket: StringPublicKeySchema,
-  lastTransactedAt: StringIntSchema,
-  mathCounter: StringIntSchema,
-  pairState: z.nativeEnum(PairState),
-  validation: ValidationPairSchema,
-
-  fundsInCurrentEpoch: StringIntSchema,
-  fundsInNextEpoch: StringIntSchema,
-  lastCalculatedSlot: StringIntSchema,
-  lastCalculatedTimestamp: StringIntSchema,
-  rewardsToHarvest: StringIntSchema,
-  rewardsToHarvested: StringIntSchema,
-})
-
-export const BondTradeTransactionSchema = z.object({
-  publicKey: StringPublicKeySchema,
-  amountOfBonds: StringIntSchema,
-  bondOffer: StringPublicKeySchema,
-  bondTradeTransactionState: z.nativeEnum(BondTradeTransactionV2State),
-  bondTradeTransactionType: z.nativeEnum(BondTradeTransactionV2Type),
-  borrowerFullRepaidAmount: StringIntSchema,
-  borrowerOriginalLent: StringIntSchema,
-  repaymentCallAmount: StringIntSchema, //? Stores value that borrower needs to pay (NOT value that lender receives)
-  currentRemainingLent: StringIntSchema,
-  fbondTokenMint: StringPublicKeySchema,
-  feeAmount: StringIntSchema,
-  interestSnapshot: StringIntSchema,
-  isDirectSell: z.boolean(),
-  lenderFullRepaidAmount: StringIntSchema,
-  lenderOriginalLent: StringIntSchema,
-  lendingToken: z.nativeEnum(LendingTokenType),
-  partialRepaySnapshot: StringIntSchema,
-  redeemResult: z.nativeEnum(RedeemResult),
-  redeemedAt: StringIntSchema,
-  repayDestination: z.nativeEnum(RepayDestination),
-  seller: StringPublicKeySchema,
-  solAmount: StringIntSchema,
-  soldAt: StringIntSchema,
-  terminationFreeze: StringIntSchema,
-  terminationStartedAt: StringIntSchema,
-  user: StringPublicKeySchema,
-})
-
-export const FraktBondSchema = z.object({
-  publicKey: StringPublicKeySchema,
-  activatedAt: StringIntSchema,
-  actualReturnedAmount: StringIntSchema,
-  amountToReturn: StringIntSchema,
-  banxStake: StringPublicKeySchema,
-  bondTradeTransactionsCounter: z.number(),
-  borrowedAmount: StringIntSchema,
-  currentPerpetualBorrowed: StringIntSchema,
-  fbondIssuer: StringPublicKeySchema,
-  fbondTokenMint: StringPublicKeySchema,
-  fbondTokenSupply: StringIntSchema,
-  fraktBondState: z.nativeEnum(FraktBondState),
-  fraktMarket: StringPublicKeySchema,
-  lastTransactedAt: StringIntSchema,
-  liquidatingAt: StringIntSchema,
-  refinanceAuctionStartedAt: StringIntSchema,
-  repaidOrLiquidatedAt: StringIntSchema,
-  terminatedCounter: z.number(),
-  hadoMarket: StringPublicKeySchema.optional(), //? Why optional
-})
 
 export const LoanSchema = z.object({
   publicKey: StringPublicKeySchema,
